requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==3.2.23
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.20.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
Pillow==8.4.0
PyJWT==2.3.0
pymongo==3.12.3
python-decouple==3.6
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sqlparse==0.2.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
